Move most of the remaining generators code into `cedar-policy-generators` (#67)
* builds, checkpoint 1

* builds, checkpoint 2

* cargo fmt

* nits

* more cargo fmt

* update CI for new crate

* checkpoint, builds

* checkpoint, builds

* move schema file

* cargo fmt

* cargo fmt

* Do not build code in the comment

---------
