+++ cedar-drt/fuzz/fuzz_targets/abac-type-directed.rs
@@ -19,7 +19,9 @@ use cedar_drt::*;
 use cedar_drt_inner::*;
 use cedar_policy_core::ast;
 use cedar_policy_core::entities::{Entities, TCComputation};
+use cedar_policy_generators::abac::{ABACPolicy, ABACRequest, ABACSettings};
 use cedar_policy_generators::err::Error;
+use cedar_policy_generators::schema::Schema;
 use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::{debug, info};
 use std::convert::TryFrom;
