+++ cedar-drt/fuzz/fuzz_targets/abac.rs
@@ -19,7 +19,9 @@ use cedar_drt::*;
 use cedar_drt_inner::*;
 use cedar_policy_core::ast;
 use cedar_policy_core::entities::Entities;
+use cedar_policy_generators::abac::{ABACPolicy, ABACRequest, ABACSettings};
 use cedar_policy_generators::hierarchy::Hierarchy;
+use cedar_policy_generators::schema::Schema;
 use cedar_policy_validator::{
     ActionBehavior, ValidationMode, Validator, ValidatorNamespaceDef, ValidatorSchemaFragment,
 };
