+++ cedar-drt/fuzz/fuzz_targets/partial-eval.rs
@@ -24,7 +24,9 @@ use cedar_policy_core::ast::PolicySet;
 use cedar_policy_core::authorizer::Authorizer;
 use cedar_policy_core::entities::{Entities, TCComputation};
 pub use cedar_policy_core::*;
+use cedar_policy_generators::abac::{ABACPolicy, ABACRequest, ABACSettings};
 use cedar_policy_generators::err::Error;
+use cedar_policy_generators::schema::Schema;
 use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::debug;
 use smol_str::SmolStr;
