+++ cedar-drt/fuzz/fuzz_targets/pp.rs
@@ -17,9 +17,11 @@
 #![no_main]
 
 use cedar_drt::initialize_log;
-use cedar_drt_inner::{fuzz_target, ABACPolicy, ABACSettings, Schema};
+use cedar_drt_inner::fuzz_target;
 use cedar_policy_core::ast::{EntityType, ExprKind, Literal, StaticPolicy, Template};
-use cedar_policy_core::parser::{err, parse_policy};
+use cedar_policy_core::parser::{self, parse_policy};
+use cedar_policy_generators::abac::{ABACPolicy, ABACSettings};
+use cedar_policy_generators::schema::Schema;
 use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::debug;
 
@@ -67,7 +69,7 @@ fn contains_unspecified_entities(p: &StaticPolicy) -> bool {
 
 // round-tripping of a policy
 // i.e., print a policy to string and parse it back
-fn round_trip(p: &StaticPolicy) -> Result<StaticPolicy, err::ParseErrors> {
+fn round_trip(p: &StaticPolicy) -> Result<StaticPolicy, parser::err::ParseErrors> {
     parse_policy(None, &p.to_string())
 }
