+++ cedar-drt/fuzz/fuzz_targets/rbac.rs
@@ -20,6 +20,8 @@ use cedar_drt_inner::*;
 use cedar_policy_core::ast;
 use cedar_policy_core::entities::Entities;
 use cedar_policy_generators::err::Result;
+use cedar_policy_generators::policy::GeneratedLinkedPolicy;
+use cedar_policy_generators::rbac::{RBACHierarchy, RBACPolicy, RBACRequest};
 use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::info;
 use std::convert::TryFrom;
