+++ cedar-drt/fuzz/fuzz_targets/validation-drt-type-directed.rs
@@ -18,6 +18,8 @@
 use cedar_drt::*;
 use cedar_drt_inner::*;
 use cedar_policy_core::ast;
+use cedar_policy_generators::abac::{ABACPolicy, ABACSettings};
+use cedar_policy_generators::schema::Schema;
 use cedar_policy_validator::{ActionBehavior, ValidatorNamespaceDef, ValidatorSchemaFragment};
 use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::{debug, info};
