+++ cedar-drt/fuzz/fuzz_targets/validation-pbt.rs
@@ -20,8 +20,10 @@ use cedar_drt_inner::*;
 use cedar_policy_core::ast;
 use cedar_policy_core::authorizer::Authorizer;
 use cedar_policy_core::entities::Entities;
+use cedar_policy_generators::abac::{ABACPolicy, ABACRequest, ABACSettings};
 use cedar_policy_generators::err::{Error, Result};
 use cedar_policy_generators::hierarchy::Hierarchy;
+use cedar_policy_generators::schema::Schema;
 use cedar_policy_validator::{
     ActionBehavior, ApplySpec, NamespaceDefinition, ValidationMode, Validator,
     ValidatorNamespaceDef, ValidatorSchemaFragment,
