+++ cedar-policy-generators/Cargo.toml
@@ -10,3 +10,4 @@ cedar-policy-core = { path = "../cedar/cedar-policy-core", version = "2.*", feat
 cedar-policy-validator = { path = "../cedar/cedar-policy-validator", version = "2.*", features = ["arbitrary"] }
 highway = "0.8.1"
 serde = { version = "1.0", features = ["derive"] }
+smol_str = "0.2"
