+++ cedar-policy-generators/src/schema.rs
@@ -14,25 +14,23 @@
  * limitations under the License.
  */
 
-use super::abac::{
+use crate::abac::{
     ABACPolicy, ABACRequest, ABACSettings, AttrValue, AvailableExtensionFunctions, ConstantPool,
     Type, UnknownPool,
 };
-use super::{ActionConstraint, PrincipalOrResourceConstraint};
-use crate::{gen, uniform};
-use ast::{Effect, PolicyID};
-use cedar_policy_core::ast;
-use cedar_policy_generators::collections::{HashMap, HashSet};
-use cedar_policy_generators::err::{while_doing, Error, Result};
-use cedar_policy_generators::hierarchy::Hierarchy;
-use cedar_policy_generators::size_hint_utils::{
-    size_hint_for_choose, size_hint_for_range, size_hint_for_ratio,
-};
+use crate::collections::{HashMap, HashSet};
+use crate::err::{while_doing, Error, Result};
+use crate::hierarchy::Hierarchy;
+use crate::policy::{ActionConstraint, GeneratedPolicy, PrincipalOrResourceConstraint};
+use crate::request::Request;
+use crate::size_hint_utils::{size_hint_for_choose, size_hint_for_range, size_hint_for_ratio};
+use crate::{accum, gen, gen_inner, uniform};
+use arbitrary::{self, Arbitrary, Unstructured};
+use cedar_policy_core::ast::{self, Effect, PolicyID};
 use cedar_policy_validator::{
     ActionType, ApplySpec, AttributesOrContext, EntityType, NamespaceDefinition, SchemaFragment,
     TypeOfAttribute,
 };
-use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use smol_str::SmolStr;
 use std::collections::BTreeMap;
 use std::str::FromStr;
@@ -270,7 +268,7 @@ fn uid_for_action_name(namespace: Option<SmolStr>, action_name: &SmolStr) -> ast
     let namespace_prefix = namespace.map(|ns| format!("{ns}::")).unwrap_or_default();
     format!("{}Action::\"{}\"", namespace_prefix, action_name)
                 .parse()
-                .unwrap_or_else(|e| panic!("schema actions should all be valid EntityUIDs in this context, but {:?} led to an invalid one: {}", action_name, e))
+                .unwrap_or_else(|e| panic!("schema actions should all be valid EntityUIDs in this context, but {action_name:?} led to an invalid one: {e}"))
 }
 
 /// internal helper function, convert a SchemaType to a Type (loses some
@@ -388,14 +386,14 @@ fn unwrap_attrs_or_context(
 fn build_qualified_entity_type(namespace: Option<SmolStr>, name: Option<&str>) -> ast::EntityType {
     match name {
         Some(name) => {
-            let type_id: ast::Id = name.parse().unwrap_or_else(|_| {
-                panic!("Valid name required to build entity type. Got {}", name)
-            });
+            let type_id: ast::Id = name
+                .parse()
+                .unwrap_or_else(|_| panic!("Valid name required to build entity type. Got {name}"));
             let type_namespace: Option<ast::Name> = match namespace.as_deref() {
                 None => None,
                 Some("") => None, // we consider "" to be the same as the empty namespace
                 Some(ns) => Some(ast::Name::from_str(&ns).unwrap_or_else(|_| {
-                    panic!("Valid namespace required to build entity type. Got {}", ns)
+                    panic!("Valid namespace required to build entity type. Got {ns}")
                 })),
             };
             match type_namespace {
@@ -651,6 +649,7 @@ impl Schema {
             attributes_by_type,
         })
     }
+    /// size hint for arbitrary()
     pub fn arbitrary_size_hint(depth: usize) -> (usize, Option<usize>) {
         arbitrary::size_hint::and_all(&[
             <HashSet<ast::Name> as Arbitrary>::size_hint(depth),
@@ -791,6 +790,7 @@ impl Schema {
             .collect::<Result<_>>()?;
         Ok(hierarchy_no_attrs.replace_entities(entities))
     }
+    /// size hint for arbitrary_hierarchy()
     pub fn arbitrary_hierarchy_size_hint(_depth: usize) -> (usize, Option<usize>) {
         (0, None)
     }
@@ -810,6 +810,7 @@ impl Schema {
             u,
         )
     }
+    /// size hint for arbitrary_principal_uid()
     pub fn arbitrary_principal_uid_size_hint(depth: usize) -> (usize, Option<usize>) {
         arbitrary::size_hint::and(
             size_hint_for_choose(None),
@@ -827,6 +828,7 @@ impl Schema {
             .map_err(|e| while_doing("choosing an action", e))?;
         Ok(uid_for_action_name(self.namespace.clone(), action))
     }
+    /// size hint for arbitrary_action_uid()
     pub fn arbitrary_action_uid_size_hint(_depth: usize) -> (usize, Option<usize>) {
         size_hint_for_choose(None)
     }
@@ -845,6 +847,7 @@ impl Schema {
             u,
         )
     }
+    /// size hint for arbitrary_resource_uid()
     pub fn arbitrary_resource_uid_size_hint(depth: usize) -> (usize, Option<usize>) {
         arbitrary::size_hint::and(
             size_hint_for_choose(None),
@@ -883,6 +886,7 @@ impl Schema {
             Some(hierarchy) => hierarchy.arbitrary_uid_with_type(ty, u),
         }
     }
+    /// size hint for arbitrary_uid_with_type()
     fn arbitrary_uid_with_type_size_hint(depth: usize) -> (usize, Option<usize>) {
         arbitrary::size_hint::or(
             <ast::Eid as Arbitrary>::size_hint(depth),
@@ -1061,6 +1065,7 @@ impl Schema {
         }
     }
 
+    /// generate an arbitrary `Value` of the given `target_type`
     pub fn arbitrary_value_for_type(
         &self,
         target_type: &Type,
@@ -1812,9 +1817,8 @@ impl Schema {
         u: &mut Unstructured<'_>,
     ) -> Result<ast::Expr> {
         if self.should_generate_unknown(max_depth, u)? {
-            let name = self
-                .unknown_pool
-                .alloc(target_type.clone(), self, hierarchy, u)?;
+            let v = self.arbitrary_value_for_type(&target_type, hierarchy, 3, u)?;
+            let name = self.unknown_pool.alloc(target_type.clone(), v);
             Ok(ast::Expr::unknown(name))
         } else {
             match target_type {
@@ -3190,7 +3194,7 @@ impl Schema {
         u: &mut Unstructured<'_>,
     ) -> Result<ABACPolicy> {
         let id = u.arbitrary()?;
-        let annotations = u.arbitrary()?;
+        let annotations: HashMap<ast::Id, String> = u.arbitrary()?;
         let effect = u.arbitrary()?;
         let principal_constraint = self.arbitrary_principal_constraint(hierarchy, u)?;
         let action_constraint = self.arbitrary_action_constraint(u, Some(3))?;
@@ -3217,15 +3221,15 @@ impl Schema {
         for constraint in abac_constraints {
             conjunction = ast::Expr::and(conjunction, constraint);
         }
-        Ok(ABACPolicy(super::GeneratedPolicy {
+        Ok(ABACPolicy(GeneratedPolicy::new(
             id,
             annotations,
             effect,
             principal_constraint,
             action_constraint,
             resource_constraint,
-            abac_constraints: conjunction,
-        }))
+            conjunction,
+        )))
     }
 
     /// size hint for arbitrary_policy()
@@ -3326,6 +3330,7 @@ impl Schema {
         )
     }
 
+    /// generate an arbitrary `ABACRequest` conforming to the schema
     pub fn arbitrary_request(
         &self,
         hierarchy: &Hierarchy,
@@ -3342,7 +3347,7 @@ impl Schema {
             )
             .expect("Failed to select action from map.");
         // now generate a valid request for that Action
-        Ok(ABACRequest(super::Request {
+        Ok(ABACRequest(Request {
             principal: match action
                 .applies_to
                 .as_ref()
@@ -3401,10 +3406,12 @@ impl Schema {
             },
         }))
     }
+    /// size hint for arbitrary_request()
     pub fn arbitrary_request_size_hint(_depth: usize) -> (usize, Option<usize>) {
         arbitrary::size_hint::and(size_hint_for_choose(None), (1, None))
     }
 
+    /// Get the namespace of this `Schema`, if any
     pub fn namespace(&self) -> &Option<SmolStr> {
         &self.namespace
     }
